requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
beautifulsoup4==4.9.3
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-bootstrap4==2.3.1
django-crispy-forms==1.11.0
django-filter==2.4.0
django-formtools==2.2
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
future==0.18.2
html5lib==1.1
importlib-metadata==2.1.1
Pillow==8.1.0
pipi==1.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.5
reportlab==3.5.59
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
zipp==3.4.0
